+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 # Component Enhancements
 
+* `Table` now lets you add an `aria-describedby` attribute via the `ariaDescribedBy` prop.
 * `TableAjax` component now uses the data-state attribute and `aria-busy`.
 
 # 1.4.6